+++ packages/overture-schema-codegen/tests/conftest.py
@@ -1,6 +1,7 @@
 """Shared pytest fixtures for overture-schema-codegen tests."""
 
 import pytest
+from codegen_test_support import find_model_class
 from overture.schema.codegen.model_extraction import extract_model
 from overture.schema.codegen.specs import ModelSpec
 from overture.schema.core.discovery import discover_models
@@ -21,16 +22,6 @@ def update_golden(request: pytest.FixtureRequest) -> bool:
     return bool(request.config.getoption("--update-golden"))
 
 
-def _find_model_class(name: str, models: dict[object, object]) -> type[BaseModel]:
-    """Find a discovered model class by name."""
-    matches = [v for v in models.values() if getattr(v, "__name__", None) == name]
-    assert matches, f"{name} model not found"
-    match = matches[0]
-    assert isinstance(match, type)
-    assert issubclass(match, BaseModel)
-    return match
-
-
 @pytest.fixture
 def all_discovered_models() -> dict:
     """Discover and return all registered Overture models."""
@@ -40,7 +31,7 @@ def all_discovered_models() -> dict:
 @pytest.fixture
 def building_class(all_discovered_models: dict) -> type[BaseModel]:
     """Get the Building model class."""
-    return _find_model_class("Building", all_discovered_models)
+    return find_model_class("Building", all_discovered_models)
 
 
 @pytest.fixture
@@ -52,10 +43,10 @@ def building_spec(building_class: type[BaseModel]) -> ModelSpec:
 @pytest.fixture
 def place_class(all_discovered_models: dict) -> type[BaseModel]:
     """Get the Place model class."""
-    return _find_model_class("Place", all_discovered_models)
+    return find_model_class("Place", all_discovered_models)
 
 
 @pytest.fixture
 def division_class(all_discovered_models: dict) -> type[BaseModel]:
     """Get the Division model class."""
-    return _find_model_class("Division", all_discovered_models)
+    return find_model_class("Division", all_discovered_models)
