+++ packages/overture-schema-transportation-theme/pyproject.toml
@@ -27,3 +27,86 @@ packages = ["src/overture"]
 [project.entry-points."overture.models"]
 "overture:transportation:connector" = "overture.schema.transportation:Connector"
 "overture:transportation:segment" = "overture.schema.transportation:Segment"
+ 
+[[examples.Connector]]
+id = "39542bee-230f-4b91-b7e5-a9b58e0c59b1"
+geometry = "POINT (-176.5472979 -43.9679472)"
+version = 1
+theme = "transportation"
+type = "connector"
+
+[examples.Connector.bbox]
+xmin = -176.54730224609375
+xmax = -176.54727172851562
+ymin = -43.96794891357422
+ymax = -43.96794128417969
+
+[[examples.Connector.sources]]
+property = ""
+dataset = "OpenStreetMap"
+record_id = "null"
+update_time = "null"
+confidence = "null"
+between = "null"
+
+[[examples.Segment]]
+id = "1bc62f3b-08b5-42b8-89fe-36f685f60455"
+geometry = "LINESTRING (-176.5636191 -43.954404, -176.5643637 -43.9538145, -176.5647264 -43.9535274, -176.5649947 -43.953251)"
+version = 1
+subtype = "road"
+class = "residential"
+routes = "null"
+subclass_rules = "null"
+access_restrictions = "null"
+level_rules = "null"
+destinations = "null"
+prohibited_transitions = "null"
+road_flags = "null"
+speed_limits = "null"
+width_rules = "null"
+subclass = "null"
+rail_flags = "null"
+theme = "transportation"
+type = "segment"
+
+[examples.Segment.bbox]
+xmin = -176.5650177001953
+xmax = -176.56361389160156
+ymin = -43.954410552978516
+ymax = -43.953250885009766
+
+[[examples.Segment.sources]]
+property = ""
+dataset = "OpenStreetMap"
+record_id = "w53435546@6"
+update_time = "2021-05-03T06:37:03Z"
+confidence = "null"
+between = "null"
+
+[examples.Segment.names]
+primary = "Meteorological Lane"
+common = "null"
+
+[[examples.Segment.names.rules]]
+variant = "common"
+language = "null"
+perspectives = "null"
+value = "Meteorological Lane"
+between = "null"
+side = "null"
+
+[[examples.Segment.connectors]]
+connector_id = "15b2c131-9137-4add-88c6-2acd3fa61355"
+at = 0.0
+
+[[examples.Segment.connectors]]
+connector_id = "23ae2702-ef77-4d2e-b39d-77360b696d20"
+at = 0.523536154
+
+[[examples.Segment.connectors]]
+connector_id = "8e944ce1-4b81-49eb-a823-7d98779c855c"
+at = 1.0
+
+[[examples.Segment.road_surface]]
+value = "gravel"
+between = "null"
